requirements.txt
faiss-cpu
python-liquid
langchain
tqdm
transformers
sentence_transformers
pyserini
tiktoken
datasets
accelerate
textgrad
